requirements.txt
# Core ETL and blockchain dependencies
web3==6.11.3                    # Ethereum web3 client (equivalent to org.web3j:core)
pyyaml==6.0.1                   # YAML configuration parsing
python-dotenv==1.0.0            # Environment variable management
asyncio-mqtt==0.16.1            # Async event handling

# Event handling and messaging
pydantic==2.5.0                 # Data validation and serialization
typing-extensions==4.8.0        # Type hints support

# Logging and utilities
loguru==0.7.2                   # Enhanced logging (alternative to Spring Boot logging)
click==8.1.7                    # CLI interface for ETL scripts

# Data processing
pandas==2.1.4                   # Data manipulation and analysis
numpy==1.24.3                   # Numerical computing

# HTTP and API clients
requests==2.31.0                # HTTP requests
aiohttp==3.9.1                  # Async HTTP client
websockets==12.0                # WebSocket support for real-time events

# Development and testing
pytest==7.4.3                   # Testing framework
pytest-asyncio==0.21.1          # Async testing support
black==23.11.0                  # Code formatting
flake8==6.1.0                   # Linting

# Optional: Database support (if needed for event storage)
# sqlalchemy==2.0.23            # Database ORM
# psycopg2-binary==2.9.9        # PostgreSQL adapter